Remove OpenMPI from base image
We might re-include it later when/if we add an MPI example.

Excluded for now as it drags in gfortran that depends on gcc etc., so this commit
essentially cuts the size of the final image to half (~500mb) instead of 1gb!

We think of code to be developed at the docker host machine and
the docker images to only contain run-time dependencies.

On a similar note we could consider skipping jupyterslurm and just ssh into the
user slurmjupyter node. But this is skipped for now.